requirements.txt
pyyaml>=6.0
jinja2>=3.1
# Optional: only needed if benchmarking via Anthropic API (claude runner)
# anthropic>=0.40
